requirements.txt
openai
gitpython
requests